+++ include/gern/codegen/finalizer.h
@@ -1,0 +1,37 @@
+#include "codegen/lower_visitor.h"
+#include "utils/scoped_set.h"
+
+namespace gern {
+
+class Finalizer : public LowerIRVisitor {
+public:
+    Finalizer(LowerIR ir)
+        : ir(ir) {
+    }
+
+    LowerIR finalize();
+    using LowerIRVisitor::visit;
+    void visit(const AllocateNode *);
+    void visit(const FreeNode *);
+    void visit(const InsertNode *);
+    void visit(const QueryNode *);
+    void visit(const ComputeNode *);
+    void visit(const IntervalNode *);
+    void visit(const BlankNode *);
+    void visit(const DefNode *);
+    void visit(const AssertNode *);
+    void visit(const BlockNode *);
+    void visit(const GridDeclNode *);
+    void visit(const SharedMemoryDeclNode *);
+    void visit(const OpaqueCall *);
+
+    util::ScopedSet<AbstractDataTypePtr> getToFree() const;
+
+private:
+    LowerIR ir;
+    LowerIR final_ir;
+    util::ScopedSet<AbstractDataTypePtr> to_free;
+    util::ScopedSet<const InsertNode *> to_insert;
+};
+
+}  // namespace gern
